Move computeDeployID() back to singularity package.
Make TestModifyResources() variable references explicit and use new dummy scheme.

Dummy rectification client stores actual Deployable structs.
